[refactor](reader) move reader from vec/exec/scan to vec/exec/format
